python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-insights"
version = "1.0.0"
description = "Chess game analysis tool with Stockfish integration"
requires-python = ">=3.10"
dependencies = [
    "python-chess>=1.10.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
chess-insights = "chess_insights.cli:main"

[tool.setuptools.packages.find]
where = ["."]
